requirements.txt
# Core ML / CV
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Vector search
faiss-cpu>=1.7.4

# Clustering / metrics
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0

# Config
PyYAML>=6.0

# Demo app
gradio>=4.0.0

# API / Backend
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.66.0
pymongo>=4.5.0
